Merged Pull Request '#124 feature/metrics-example->main: Add two new examples'
* FEAT: Added an example to produce metrics for the IPI data file.

* FEAT: Added the average number of geographic areas to the metric.
FEAT: Added a sample percentage to balance coverage versus elapsed execution time.
PERF: Improved the performance of the example.

* FEAT: Added queries per second to the logger during processing.

* OPTIM: Set concurrency in case lower performance profiles are used.

* FEAT: Added the basis of a second example that will compare the positions returned from IPI with some ground truth provided as a CSV file.

* FEAT: improve logging, noalloc IP iteration

* CLEANUP: Usings.

* CLEANUP:

* FEAT: Added an example to work out the distance between a truth point and IP address and the latitude and longitude returned from the IP location service.

* BUG: Remove unused usings.
BUG: Don't process duplicate IP addresses in the source data.

* OPTIM: Moved the area methods into a common project to be shared between examples.

* FEAT: Added the contains property to the compare result to indicate if the area returned from IPI contains the latitude and longitude provided in the truth.
DOC: Modified comments to make clearer the purpose of the examples. These are no common examples but provided to help those evaluating IPI understand the accuracy in their environments and with their data.

* CLEANUP: Remove unneccesary casts
Simplify collection initialization.
make run static as no instance data is being used.
remove unused paramter 'ipiEngine'
Pass in cancellation tokens where expected. This isnt really necessary as we don't expect to be cancelling for a short program but its still worth including.
REF: Update Submolude to contain the geoip evidence data for testing.

* CLEANUP: Use modern C# syntax pattern matching to make the retrieval of values cleaner.

* FEAT: Use Csv writer to write metrics data in csv format

* FEAT: Modified the area calculation to use rectangles of no more than 1 unit of latitude and longitude when calculating the area to improve the accuracy of the resulting value when used with large areas.
FEAT: Some WKT values are not simple polygons. This needs to be addressed in the source data. There is a temporary exception handler added as such areas can't be handled. The fall back calculation is to take the reported area without apply the grid rectangle.

* FEAT/TEST: Add Tests for Compare & Metrics.
Introduce contstants for new geoip truth file.
OPTIM: In metrics console, move the sample filter to the producer to reduce processing and test time.
CLEANUP: Use filescoped namespace for the test class.

* CLEANUP: Remove unused usings.

* CLEANUP: Align versions to .NET 8.

* BUG: Write out records correctly in Metrics Console

* UPDATE: Use the latest versions of 51Degrees IPI pacakges to address issue with WKT files.
TEST: Modified to use other test data. Note: in rare cases the intersection still fails with what appears to be valid data.

* BUG: Remove speech marks from groups.Key and use record.Value.ToString in order to get the proper formatiing.
REVERT: put the sample percentage back in the consumers as you want to take a sample of random ips, not a sample of ranges.

* REF: Update Submodule to main that includes GeoIp Evidence file.
CLEANUP: Rename file name of comparison file.

* BUG: Update Compare-Console to no longer retreive weight list properties.
CLEANUP: Remove unused usings
REF: Resolve package conflicts, update all to latest ip-intelligence package.

* DOC: Tweak the descriptions of Metric-Console and Compare-Console

* FEAT: Allow for passing in of external logger so that automated comparison workers can be logged to.
BUG: Correct the name of the sample file used in ip-intelligence-data + use the constant. Remove TODO.

* BUG: Use the FindFile method to get the Geoip comparison file

---------

Co-authored-by: Eugene Dorfman <eugene.dorfman@gmail.com>
Co-authored-by: James Rogers <JamesR@51degrees.com>
Co-authored-by: Jamesr51d <122470673+Jamesr51d@users.noreply.github.com>